Rename dataframe.{h,cc} to data_frame.{h,cc}
