+++ src/cross_mapping.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-#include "dataframe.h"
+#include "data_frame.h"
 
 class CrossMapping
 {
