+++ src/embedding_dim.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-#include "dataframe.h"
+#include "data_frame.h"
 
 class EmbeddingDim
 {
