+++ src/nearest_neighbors_gpu.h
@@ -1,7 +1,7 @@
 #ifndef __NEAREST_NEIGHBORS_GPU_H__
 #define __NEAREST_NEIGHBORS_GPU_H__
 
-#include "dataframe.h"
+#include "data_frame.h"
 #include "lut.h"
 #include "nearest_neighbors.h"
