+++ src/stats.h
@@ -1,7 +1,7 @@
 #ifndef __STATS_H__
 #define __STATS_H__
 
-#include "dataframe.h"
+#include "data_frame.h"
 
 float corrcoef(const Series &x, const Series &y);
