+++ test/cross_mapping_test.cc
@@ -1,7 +1,7 @@
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 
-#include "../src/dataframe.h"
+#include "../src/data_frame.h"
 #include "../src/lut.h"
 #include "../src/nearest_neighbors_cpu.h"
 #include "../src/simplex_cpu.h"
